Wait for YouTube player to load before removing ads from it (issue 43)
This approach injects a script tag containing the ad-removal code into the page itself. This allows us to check whether the Flash object is loaded by trying to call a method on it. If the method exists, the object is loaded and it's more likely that replacing it with a modified duplicate will work.

Even if this doesn't help issue 43, it at least opens the door to directly messing with YouTube's scripts in the future.

--HG--
extra : convert_revision : ab25cea1b6c0f750be03481ae23572e951eef247