Don't explicitly hide elements that should be hidden by elemhide stylesheet (issue 229).
The call to hideBySelectorStrings() was necessitated by the elemhide stylesheet not always working properly. Now I think that was because of misuse of the title attribute (issue 41, now fixed). Removing this call speeds up page rendering noticeably on sites like engadget.com because we don't have to call querySelectorAll for all those selectors.